pyproject.toml
[tool.poetry]
name = "contacts-keeper"
version = "0.1.0"
description = ""
authors = ["lrayne <legaterayne@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
black = "^24.4.0"
mysql-connector = "^2.2.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
skip-string-normalization = true
